+++ pittgoogle/registry.py
@@ -4,18 +4,17 @@ import importlib.resources
 import logging
 from typing import Final
 
+import attrs
 import yaml
-from attrs import define
 
-from . import types_
-from .exceptions import SchemaNotFoundError
+from . import exceptions, schema
 
 LOGGER = logging.getLogger(__name__)
 PACKAGE_DIR = importlib.resources.files(__package__)
 SCHEMA_MANIFEST = yaml.safe_load((PACKAGE_DIR / "registry_manifests/schemas.yml").read_text())
 
 
-@define(frozen=True)
+@attrs.define(frozen=True)
 class ProjectIds:
     """Registry of Google Cloud Project IDs."""
 
@@ -32,7 +31,7 @@ class ProjectIds:
     """Project running classifiers for ELAsTiCC alerts and reporting to DESC."""
 
 
-@define(frozen=True)
+@attrs.define(frozen=True)
 class Schemas:
     """Registry of schemas used by Pitt-Google.
 
@@ -55,7 +54,7 @@ class Schemas:
     """
 
     @staticmethod
-    def get(schema_name: str) -> types_.Schema:
+    def get(schema_name: str | None) -> schema.Schema:
         """Return the schema with name matching `schema_name`.
 
         Returns:
@@ -63,15 +62,23 @@ class Schemas:
                 Schema from the registry with name matching `schema_name`.
 
         Raises:
-            SchemaNotFoundError:
+            SchemaError:
                 If a schema with name matching `schema_name` is not found in the registry.
-            SchemaNotFoundError:
+            SchemaError:
                 If a schema definition cannot be loaded but one will be required to read the alert bytes.
         """
+        # If no schema_name provided, return the default.
+        if schema_name is None:
+            LOGGER.warning("No schema name provided. Returning a default schema.")
+            mft_schema = [
+                schema for schema in SCHEMA_MANIFEST if schema["name"] == "default_schema"
+            ][0]
+            return schema.Schema._from_yaml(schema_dict=mft_schema)
+
         # Return the schema with name == schema_name, if one exists.
         for mft_schema in SCHEMA_MANIFEST:
             if mft_schema["name"] == schema_name:
-                return types_.Schema._from_yaml(schema_dict=mft_schema)
+                return schema.Schema._from_yaml(schema_dict=mft_schema)
 
         # Return the schema with name ~= schema_name, if one exists.
         for mft_schema in SCHEMA_MANIFEST:
@@ -79,10 +86,10 @@ class Schemas:
             # Catches names like 'lsst.v<MAJOR>_<MINOR>.alert' where users replace '<..>' with custom values.
             split_name, split_mft_name = schema_name.split("."), mft_schema["name"].split(".")
             if all([split_mft_name[i] == split_name[i] for i in [0, -1]]):
-                return types_.Schema._from_yaml(schema_dict=mft_schema, name=schema_name)
+                return schema.Schema._from_yaml(schema_dict=mft_schema, name=schema_name)
 
         # That's all we know how to check so far.
-        raise SchemaNotFoundError(
+        raise exceptions.SchemaError(
             f"{schema_name} not found. For valid names, see `pittgoogle.Schemas.names`."
         )